Moved invations to a separe interface.
